move rest of the code from downloader to utils
