+++ crashsimilarity/top_similar_traces.py
@@ -1,8 +1,9 @@
 # CLI INTERFACE THAT TAKES STACK TRACE AS INPUT AND RETURNS SIMILAR STACK TRACES
-import download_data
-import crash_similarity
+from crashsimilarity import utils, crash_similarity
 import argparse
 import sys
+
+from crashsimilarity.downloader import SocorroDownloader
 
 
 def parse_args(args):
@@ -16,8 +17,8 @@ def parse_args(args):
 if __name__ == '__main__':
     args = parse_args(sys.argv[1:])
     # downloads some data (e.g. the past 7 days)
-    download_data.download_crashes(days=7, product=args.product)
-    paths = download_data.get_paths(days=7, product=args.product)
+    utils.download_crashes(days=7, product=args.product)
+    paths = utils.get_paths(days=7, product=args.product)
 
     # reads the corpus
     corpus = crash_similarity.read_corpus(paths)
@@ -26,7 +27,7 @@ if __name__ == '__main__':
     model = crash_similarity.train_model(corpus)
 
     # gets the stack_trace corresponding to the crash_id (input by the user as an argument)
-    stack_trace = download_data.download_stack_trace_for_crashid(args.crash_id)
+    stack_trace = SocorroDownloader().download_crash(args.crash_id)['proto_signature']
 
     # returns the top similar stack traces (number of stack traces returned = args.top)
     similarities = crash_similarity.top_similar_traces(model, corpus, stack_trace, args.top)
